[Finder] Fixed support for Streams (eg. Phar archives)
The 'AbstractFindAdapter' does now an validation of the realpath result. In case
of errors an 'AdapterFailureException' will be thrown.

Modified the 'AdapterInterface' to enable path validation for the 'isSupported'
method which will be called by the user. Therefore a new parameter was added.
Accordingly to this new feature the 'AbstractAdapter' implementation was reworked
to run an newly added method 'canBeUsedOnPath' which does an adapter-based validation
of the given path. Additionally the cache mechanism of the method 'isSupported' of
the class 'AbstractAdapter' was extended. All adapters received an implementation of the
path validation method 'canBeUsedOnPath'.

Added two testcases which prove the adapter abilities to search in phar archives
and validate the support of handling streams. The helper method 'toAbsoluteFixtures'
of 'FinderTest' was extended to recognize phar archives and prefix them with the
phar scheme. This method converts now the directory separators for an given
path to provide compatibility to Windows systems.